web-app/requirements.txt
Django
psycopg2
requests
awscli
requests-aws4auth
Pillow
django-environ
django_bootstrap5
django-extensions
whitenoise